+++ src/components/widgets/WidgetKeyResultDetails.vue
@@ -113,7 +113,7 @@
 import { mapState } from 'vuex';
 import { db } from '@/config/firebaseConfig';
 import { dateLong } from '@/util';
-import formattedPeriod from '@/util/okr';
+import { formattedPeriod } from '@/util/okr';
 import ProfileModal from '@/components/modals/ProfileModal.vue';
 import Widget from './WidgetWrapper.vue';